requirements.txt
torch
timm
numpy
scipy
mne
PyYAML
tqdm
